centralize tap commands into one file
